pyproject.toml
[project]
name = "faq-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.118.0",
    "numpy>=1.24.4",
    "openai>=1.109.1",
    "pandas>=2.0.3",
    "python-dotenv>=1.0.1",
    "rank-bm25>=0.2.2",
    "uvicorn>=0.25.0",
    "pytest>=8.3.2",
]
